+++ mm2src/mm2_main/tests/mm2_tests/tendermint_tests.rs
@@ -8,12 +8,12 @@ use mm2_test_helpers::for_tests::{atom_testnet_conf, disable_coin, disable_coin_
 use mm2_test_helpers::structs::{RpcV2Response, TendermintActivationResult, TransactionDetails};
 use serde_json::{self as json, json};
 
-const IRIS_TEST_SEED: &str = "iris test seed";
 const ATOM_TEST_BALANCE_SEED: &str = "atom test seed";
 const ATOM_TEST_WITHDRAW_SEED: &str = "atom test withdraw seed";
 const ATOM_TICKER: &str = "ATOM";
 const ATOM_TENDERMINT_RPC_URLS: &[&str] = &["https://rpc.sentry-02.theta-testnet.polypore.xyz"];
 
+const IRIS_TEST_SEED: &str = "iris test seed";
 const IRIS_TESTNET_RPC_URLS: &[&str] = &["http://34.80.202.172:26657"];
 
 #[test]
@@ -352,12 +352,60 @@ fn test_disable_tendermint_platform_coin_with_token() {
     let activation_res = block_on(enable_tendermint_token(&mm, token));
     assert!(&activation_res.get("result").unwrap().get("balances").is_some());
 
-    // Try to disable platform coin, IRIS-TEST. This should fail due to the dependent tokens.
-    let error = block_on(disable_coin_err(&mm, "IRIS-TEST"));
-    assert_eq!(error.dependent_tokens, ["IRIS-NIMDA"]);
+    // Try to passive platform coin, IRIS-TEST.
+    let res = block_on(disable_coin(&mm, "IRIS-TEST", false));
+    assert!(res.passivized);
 
-    // Try to disable IRIS-NIMDA token first.
-    block_on(disable_coin(&mm, "IRIS-NIMDA"));
-    // Then try to disable IRIS-TEST platform coin.
-    block_on(disable_coin(&mm, "IRIS-TEST"));
+    // Try to disable IRIS-NIMDA token when platform coin is passived.
+    // This should work, because platform coin is still in the memory.
+    let res = block_on(disable_coin(&mm, "IRIS-NIMDA", false));
+    assert!(!res.passivized);
+
+    // Then try to force disable IRIS-TEST platform coin.
+    let res = block_on(disable_coin(&mm, "IRIS-TEST", true));
+    assert!(!res.passivized);
+}
+
+#[test]
+fn test_passive_coin_and_force_disable() {
+    const TEST_SEED: &str = "iris test seed";
+    let coins = json!([iris_testnet_conf(), iris_nimda_testnet_conf()]);
+    let platform_coin = coins[0]["coin"].as_str().unwrap();
+    let token = coins[1]["coin"].as_str().unwrap();
+
+    let conf = Mm2TestConf::seednode(TEST_SEED, &coins);
+    let mm = MarketMakerIt::start(conf.conf, conf.rpc_password, None).unwrap();
+
+    // Enable platform coin IRIS-TEST
+    let activation_res = block_on(enable_tendermint(&mm, platform_coin, &[], IRIS_TESTNET_RPC_URLS, false));
+    assert!(&activation_res.get("result").unwrap().get("address").is_some());
+
+    // Enable platform coin token IRIS-NIMDA
+    let activation_res = block_on(enable_tendermint_token(&mm, token));
+    assert!(&activation_res.get("result").unwrap().get("balances").is_some());
+
+    // Try to passive platform coin, IRIS-TEST.
+    let res = block_on(disable_coin(&mm, "IRIS-TEST", false));
+    assert!(res.passivized);
+
+    // Try to disable IRIS-NIMDA token when platform coin is passived.
+    // This should work, because platform coin is still in the memory.
+    let res = block_on(disable_coin(&mm, "IRIS-NIMDA", false));
+    assert!(!res.passivized);
+
+    // Re-activate passive coin
+    let activation_res = block_on(enable_tendermint(&mm, platform_coin, &[], IRIS_TESTNET_RPC_URLS, false));
+    assert!(&activation_res.get("result").unwrap().get("address").is_some());
+
+    // Enable platform coin token IRIS-NIMDA
+    let activation_res = block_on(enable_tendermint_token(&mm, token));
+    assert!(&activation_res.get("result").unwrap().get("balances").is_some());
+
+    // Try to force disable platform coin, IRIS-TEST.
+    let res = block_on(disable_coin(&mm, "IRIS-TEST", true));
+    assert!(!res.passivized);
+
+    // Try to disable IRIS-NIMDA token when platform coin force disabled.
+    // This should failed, because platform coin was purged with it's tokens.
+    block_on(disable_coin_err(&mm, "IRIS-NIMDA", false));
 }